refactor(mesh): extract projection_axes helper shared by polygon + cdt (issue 799)
Lift the 17-line dominant-axis pick that picked the 2D-projection axis
pair on a Plane3 into crate::plane as pub(crate) Axis +
projection_axes. polygon::projected_signed_area and
tessellate::cdt::triangulate now share the single helper instead of
maintaining structurally identical copies.
